requirements.txt
faust-streaming
kafka-python
psycopg2
numpy